Refactored math util to plain funcs
